+++ packages/hoc/src/__tests__/mutations/index.test.tsx
@@ -39,7 +39,7 @@ describe('graphql(mutation)', () => {
   let client: ApolloClient<NormalizedCacheObject>;
   beforeEach(() => {
     error = console.error;
-    console.error = jest.fn(() => {});
+    console.error = jest.fn(() => { });
     client = createClient(expectedData, query);
   });
 
@@ -88,7 +88,7 @@ describe('graphql(mutation)', () => {
   });
 
   it('binds a mutation to props with a custom name', () => {
-    interface Props {}
+    interface Props { };
 
     type InjectedProps = {
       customMutation: any;
@@ -243,7 +243,7 @@ describe('graphql(mutation)', () => {
       second: 2
     });
 
-    interface Props {}
+    interface Props { }
 
     const Container = graphql<Props>(queryWithVariables, {
       options: () => ({