+++ packages/testing/src/mocks/MockedProvider.tsx
@@ -3,7 +3,6 @@ import { ApolloClient, DefaultOptions, Resolvers } from 'apollo-client';
 import { ApolloCache } from 'apollo-cache';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 import { ApolloProvider } from '@apollo/react-common';
-
 import { MockLink } from './mockLink';
 import { MockedResponse } from './types';
 
@@ -24,7 +23,7 @@ export interface MockedProviderState {
 export class MockedProvider extends React.Component<
   MockedProviderProps,
   MockedProviderState
-> {
+  > {
   public static defaultProps: MockedProviderProps = {
     addTypename: true
   };