+++ packages/common/src/context/__tests__/ApolloProvider.test.tsx
@@ -1,11 +1,10 @@
 import React, { useContext } from 'react';
-import { render, cleanup } from 'react-testing-library';
+import { render, cleanup } from '@testing-library/react';
 import ApolloClient from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 import { ApolloLink } from 'apollo-link';
 
-import { ApolloProvider } from '../ApolloProvider';
-import { getApolloContext } from '../ApolloContext';
+import { ApolloProvider, getApolloContext } from '@apollo/react-common';
 
 describe('<ApolloProvider /> Component', () => {
   afterEach(cleanup);