+++ packages/components/src/__tests__/ssr/getDataFromTree.test.tsx
@@ -3,11 +3,11 @@ import ReactDOM from 'react-dom/server';
 import ApolloClient from 'apollo-client';
 import { ApolloProvider, getApolloContext } from '@apollo/react-common';
 import { getDataFromTree } from '@apollo/react-hooks';
-import { Query } from '../../';
 import gql from 'graphql-tag';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 import { mockSingleLink } from '@apollo/react-testing';
 import { DocumentNode } from 'graphql';
+import { Query } from '@apollo/react-components';
 
 describe('SSR', () => {
   describe('`getDataFromTree`', () => {