+++ packages/hoc/src/__tests__/fragments.test.tsx
@@ -1,14 +1,12 @@
 import React from 'react';
-import { render, cleanup } from 'react-testing-library';
+import { render, cleanup } from '@testing-library/react';
 import gql from 'graphql-tag';
 import ApolloClient from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 import { mockSingleLink, stripSymbols } from '@apollo/react-testing';
 import { ApolloProvider } from '@apollo/react-common';
 import { DocumentNode } from 'graphql';
-
-import { graphql } from '../graphql';
-import { ChildProps } from '../types';
+import { graphql, ChildProps } from '@apollo/react-hoc';
 
 describe('fragments', () => {
   afterEach(cleanup);