+++ packages/hoc/src/__tests__/queries/index.test.tsx
@@ -1,20 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { render, cleanup } from 'react-testing-library';
+import { render, cleanup } from '@testing-library/react';
 import gql from 'graphql-tag';
 import ApolloClient from 'apollo-client';
 import { InMemoryCache as Cache } from 'apollo-cache-inmemory';
 import { ApolloLink } from 'apollo-link';
-import {
-  mockSingleLink,
-  stripSymbols,
-  catchAsyncError
-} from '@apollo/react-testing';
+import { mockSingleLink, stripSymbols } from '@apollo/react-testing';
 import { ApolloProvider } from '@apollo/react-common';
 import { DocumentNode } from 'graphql';
-
-import { graphql } from '../../graphql';
-import { ChildProps, DataProps } from '../../types';
+import { graphql, ChildProps, DataProps } from '@apollo/react-hoc';
 
 describe('queries', () => {
   let error: typeof console.error;
@@ -399,11 +393,13 @@ describe('queries', () => {
       class extends React.Component<ChildProps<{}, Data, Vars>> {
         componentDidUpdate() {
           const { props } = this;
-          catchAsyncError(done, () => {
+          try {
             expect(props.data!.loading).toBeFalsy();
             expect(stripSymbols(props.data!.allPeople)).toEqual(data.allPeople);
             done();
-          });
+          } catch (error) {
+            done.fail(error);
+          }
         }
         render() {
           return null;