+++ Changelog.md
@@ -34,6 +34,30 @@
 
 - TODO
 
+## vNext
+
+### Improvements
+
+- Make sure `MockedProvider` is using the proper CJS/ESM bundle, when referencing `ApolloProvider`. <br/>
+  [@jure](https://github.com/jure) in [#3029](https://github.com/apollographql/react-apollo/pull/3029).
+
+- Adjust the `ApolloContext` definition to play a bit more nicely with `React.createContext` types. <br/>
+  [@JoviDeCroock](https://github.com/JoviDeCroock) in [#3018](https://github.com/apollographql/react-apollo/pull/3018)
+
+- The result of a mutation is now made available to the wrapped component, when using the `graphql` HOC. <br/>
+  [@andycarrell](https://github.com/andycarrell) in [#3008](https://github.com/apollographql/react-apollo/pull/3008) 
+
+### Bug Fixes
+
+- Removed leftover `apollo-client@beta` peer dep. <br/>
+  [@brentertz](https://github.com/brentertz) in [#3064](https://github.com/apollographql/react-apollo/pull/3064)
+
+- Stop setting optional input to `null`, when using the `graphql` HOC. <br/>
+  [@ZhengYuTay](https://github.com/ZhengYuTay) in [#3056](https://github.com/apollographql/react-apollo/pull/3056)
+
+- Fix typescript error caused by `query` being mandatory in the `fetchMore` signature. <br/>
+  [@HsuTing](https://github.com/HsuTing) in [#3065](https://github.com/apollographql/react-apollo/pull/3065)
+
 ## 2.5.6 (2019-05-22)
 
 ### Improvements